Refactor: Split featured events logic into modules
Modularized the featured events logic in `src/components/admin/useFeaturedEventsLogic.ts` into smaller, more manageable files. This improves code organization and maintainability.
